requirements.txt
annotated-types==0.7.0
anyio==4.5.2
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.16.1
gdown==5.2.0
gensim==4.3.3
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
numpy==1.24.4
packaging==25.0
pandas==2.0.3
pluggy==1.5.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic-core==2.27.2
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
scipy==1.10.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
starlette==0.44.0
tomli==2.2.1
tqdm==4.67.1
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.21.0
wrapt==1.17.2
